--- theme/snippets/nav-close-menu-button.liquid
@@ -1,1 +1,0 @@
-<button id="nav-close-menu-button" class="bg-color-transparent"><img src="{{ 'close-icon.svg' | asset_url }}" alt="{{ 'snippets.nav.close_icon.alt' | t }}" class="MobileNav__icon" /></button>
