--- theme/snippets/nav-open-menu-button.liquid
@@ -1,1 +1,0 @@
-<button id="nav-open-menu-button" class="bg-color-transparent"><img src="{{ 'menu-icon.svg' | asset_url }}" alt="{{ 'snippets.nav.menu_icon.alt' | t }}" class="MobileNav__icon" /></button>
